Assignment/module-5 rest api/restapi2/requirements.txt
django>=4.2
djangorestframework>=3.14
requests>=2.31
django-allauth>=0.54.0
python-dotenv>=1.0.0
